+++ src/database/document/task_progress.h
@@ -18,7 +18,7 @@ namespace celeritas
 
         task_progress(int64_t id, int cfg_id);
 
-        task_progress(int64_t id, int cfg_id, int progress);
+        task_progress(int64_t id, int cfg_id, int64_t progress);
 
         [[nodiscard]] int64_t get_id() const;
 
@@ -28,9 +28,9 @@ namespace celeritas
 
         [[nodiscard]] task_status_type get_status() const;
 
-        void set_progress(int progress);
+        void set_progress(int64_t progress);
 
-        void add_progress(int progress);
+        void add_progress(int64_t progress);
 
         void set_status(task_status_type status);
 
@@ -46,7 +46,7 @@ namespace celeritas
     private:
         int64_t id_;
         int cfg_id_;
-        int progress_;
+        int64_t progress_;
         task_status_type status_;
     };
 }
