+++ src/player/task/detail/base/task_context.h
@@ -10,9 +10,9 @@ namespace celeritas
     public:
         using class_type = task_context;
 
-        task_context(config::task_event_type type, task_change_type change_type, int target_id, int value);
+        task_context(config::task_event_type type, task_change_type change_type, int target_id, int64_t value);
 
-        task_context(config::task_event_type type, task_change_type change_type, int value);
+        task_context(config::task_event_type type, task_change_type change_type, int64_t value);
 
         [[nodiscard]] config::task_event_type get_type() const;
 
@@ -20,12 +20,12 @@ namespace celeritas
 
         [[nodiscard]] int get_target_id() const;
 
-        [[nodiscard]] int get_value() const;
+        [[nodiscard]] int64_t get_value() const;
 
     private:
         config::task_event_type type_;
         task_change_type change_type_;
         int target_id_;
-        int value_;
+        int64_t value_;
     };
 }
